Refactor unit tests and add feature test
